Back-port of Plus42 variable handling, first set of changes
I probably need to start this over, and re-think all of FUNC, LNSTK,
L4STK, INDEX handling, etc. The existing data structures are OK, I
think; the way Free42 combines globals and locals into one flat
structure makes sense given the absence of directories, and I don't see
any objections to the way things are organized in Plus42, either. But
the way FUNC etc. operate needs a re-think and re-design, I think; the
current code is a mess of hack upon hack upon hack, badly structured and
virtually incomprehensible. I think rewriting that logic will lead to
code that is more compact, more readable, and more robust.

So this commit will probably end up leading nowhere, which is why I'm
removing it from master and parking it in a branch. But I'm not quite
ready to throw it away yet.